+++ randovania/games/prime1/exporter/patch_data_factory.py
@@ -1,3 +1,4 @@
+from pydoc import doc
 import typing
 from random import Random
 from typing import List, Union
@@ -11,12 +12,13 @@ from randovania.game_description.resources.resource_database import ResourceData
 from randovania.game_description.resources.resource_info import CurrentResources
 from randovania.game_description.world.area_identifier import AreaIdentifier
 from randovania.game_description.world.teleporter_node import TeleporterNode
+from randovania.game_description.world.dock_node import DockNode
 from randovania.game_description.world.pickup_node import PickupNode
 from randovania.game_description.world.world_list import WorldList
 from randovania.games.game import RandovaniaGame
 from randovania.games.prime1.exporter.hint_namer import PrimeHintNamer
 from randovania.games.prime1.layout.hint_configuration import ArtifactHintMode
-from randovania.games.prime1.layout.prime_configuration import PrimeConfiguration
+from randovania.games.prime1.layout.prime_configuration import PrimeConfiguration, RoomRandoMode
 from randovania.games.prime1.layout.prime_cosmetic_patches import PrimeCosmeticPatches
 from randovania.games.prime1.patcher import prime1_elevators, prime_items
 from randovania.generator.item_pool import pickup_creator
@@ -52,11 +54,69 @@ _STARTING_ITEM_NAME_TO_INDEX = {
     "wavebuster": "Wavebuster"
 }
 
-# "Power Suit": "PowerSuit",
-# "Combat Visor": "Combat",
-# "Unknown Item 1": "Unknown1",
-# "Health Refill": "HealthRefill",
-# "Unknown Item 2": "Unknown2",
+_DOCKS_TO_SKIP = [
+    ("Frigate Orpheon", "Air Lock", 1),
+    ("Frigate Orpheon", "Air Lock", 2),
+    ("Frigate Orpheon", "Deck Alpha Mech Shaft", 0),
+    ("Frigate Orpheon", "Deck Alpha Mech Shaft", 1),
+    ("Frigate Orpheon", "Connection Elevator to Deck Alpha", 0),
+    ("Frigate Orpheon", "Connection Elevator to Deck Alpha", 1),
+    ("Frigate Orpheon", "Biotech Research Area 2", 0),
+    ("Frigate Orpheon", "Deck Gamma Monitor Hall", 0),
+    ("Frigate Orpheon", "Connection Elevator to Deck Beta", 1),
+    ("Frigate Orpheon", "Biotech Research Area 1", 2),
+    ("Frigate Orpheon", "Biotech Research Area 1", 3),
+    ("Frigate Orpheon", "Cargo Freight Lift to Deck Gamma", 1),
+    ("Frigate Orpheon", "Cargo Freight Lift to Deck Gamma", 2),
+    ("Frigate Orpheon", "Cargo Freight Lift to Deck Gamma", 3),
+    ("Frigate Orpheon", "Subventilation Shaft Section A", 0),
+    ("Frigate Orpheon", "Subventilation Shaft Section B", 0),
+    ("Frigate Orpheon", "Subventilation Shaft Section B", 1),
+    ("Frigate Orpheon", "Main Ventilation Shaft Section A", 0),
+    ("Frigate Orpheon", "Main Ventilation Shaft Section B", 0),
+    ("Frigate Orpheon", "Main Ventilation Shaft Section C", 1),
+    ("Frigate Orpheon", "Main Ventilation Shaft Section D", 0),
+    ("Frigate Orpheon", "Main Ventilation Shaft Section E", 1),
+    ("Frigate Orpheon", "Main Ventilation Shaft Section F", 1),
+    ("Frigate Orpheon", "Reactor Core", 0),
+    ("Frigate Orpheon", "Reactor Core Entrance", 0),
+    ("Frigate Orpheon", "Reactor Core Entrance", 1),
+    ("Chozo Ruins", "Main Plaza", 4),
+    ("Chozo Ruins", "Main Plaza", 5),
+    ("Chozo Ruins", "Piston Tunnel", 0),
+    ("Chozo Ruins", "Piston Tunnel", 1),
+    ("Chozo Ruins", "Training Chamber", 1),
+    ("Chozo Ruins", "Energy Core", 0),
+    ("Chozo Ruins", "Burn Dome Access", 1),
+    ("Chozo Ruins", "Furnace", 2),
+    ("Chozo Ruins", "Crossway Access West", 1),
+    ("Phendrana Drifts", "Quarantine Cave", 2),
+    ("Phendrana Drifts", "Quarantine Monitor", 0),
+    ("Phendrana Drifts", "West Tower", 0),
+    ("Phendrana Drifts", "West Tower", 1),
+    ("Phendrana Drifts", "Phendrana's Edge", 3),
+    ("Phendrana Drifts", "Security Cave", 0),
+    ("Tallon Overworld", "Reactor Core", 0),
+    ("Tallon Overworld", "Reactor Access", 0),
+    ("Tallon Overworld", "Reactor Access", 1),
+    ("Tallon Overworld", "Cargo Freight Lift to Deck Gamma", 0),
+    ("Tallon Overworld", "Life Grove Tunnel", 1),
+    ("Tallon Overworld", "Life Grove", 0),
+    ("Magmoor Caverns", "Warrior Shrine", 1),
+    ("Magmoor Caverns", "Fiery Shores", 2),
+    ("Impact Crater", "Phazon Infusion Chamber", 0),
+    ("Impact Crater", "Subchamber One", 0),
+    ("Impact Crater", "Subchamber One", 1),
+    ("Impact Crater", "Subchamber Two", 0),
+    ("Impact Crater", "Subchamber Two", 1),
+    ("Impact Crater", "Subchamber Three", 0),
+    ("Impact Crater", "Subchamber Three", 1),
+    ("Impact Crater", "Subchamber Four", 0),
+    ("Impact Crater", "Subchamber Four", 1),
+    ("Impact Crater", "Subchamber Five", 0),
+    ("Impact Crater", "Subchamber Five", 1),
+    ("Impact Crater", "Metroid Prime Lair", 0),
+]
 
 _MODEL_MAPPING = {
     (RandovaniaGame.METROID_PRIME_ECHOES, "CombatVisor INCOMPLETE"): "Combat Visor",
@@ -234,7 +294,10 @@ class PrimePatchDataFactory(BasePatchDataFactory):
                 "transports": {},
                 "rooms": {}
             }
+
             for area in world.areas:
+                world_data[world.name]["rooms"][area.name] = {"pickups":[]}
+                
                 pickup_indices = sorted(node.pickup_index for node in area.nodes if isinstance(node, PickupNode))
                 if pickup_indices:
                     world_data[world.name]["rooms"][area.name] = {
@@ -245,6 +308,12 @@ class PrimePatchDataFactory(BasePatchDataFactory):
                             for index in pickup_indices
                         ],
                     }
+
+                if self.configuration.superheated_probability != 0:
+                    world_data[world.name]["rooms"][area.name]["superheated"] = self.rng.random() <= self.configuration.superheated_probability/1000.0
+
+                if self.configuration.submerged_probability != 0:
+                    world_data[world.name]["rooms"][area.name]["submerge"] = self.rng.random() <= self.configuration.submerged_probability/1000.0
 
                 for node in area.nodes:
                     if not isinstance(node, TeleporterNode) or not node.editable:
@@ -258,6 +327,91 @@ class PrimePatchDataFactory(BasePatchDataFactory):
                         identifier.area_location.area_name,
                     )]
                     world_data[world.name]["transports"][source_name] = target
+
+        if self.configuration.room_rando == RoomRandoMode.ONE_WAY:
+            for world in db.world_list.worlds:
+                if world.name == "End of Game":
+                    continue
+                
+                area_dock_nums = dict()
+                attached_areas = dict()
+                size_indices = dict()
+                unused_docks = list()
+
+                for area in world.areas:
+                    area_dock_nums[area.name] = list()
+                    attached_areas[area.name] = list()
+                    for node in area.nodes:
+                        if not isinstance(node, DockNode):
+                            continue
+                        index = node.extra["dock_index"]
+                        if (world.name, area.name, index) in _DOCKS_TO_SKIP:
+                            continue
+                        area_dock_nums[area.name].append(index)
+                        attached_areas[area.name].append(node.default_connection.area_name)
+                        unused_docks.append((area.name, index))
+                    size_indices[area.name] = area.extra["size_index"]
+
+                self.rng.shuffle(unused_docks)
+
+                for area in world.areas:
+                    world_data[world.name]["rooms"][area.name]["doors"] = dict()
+                    for dock_num in area_dock_nums[area.name]:
+
+                        def are_rooms_compatible(src, dest):
+                            if src is None or dest is None:
+                                return False
+
+                            # both rooms must have patchable docks
+                            if len(area_dock_nums[src]) == 0 or len(area_dock_nums[dest]) == 0:
+                                return False
+
+                            # destinations cannot be in the same room
+                            if src == dest:
+                                return False
+                            
+                            # rooms cannot be neighbors
+                            if src in attached_areas[dest]:
+                                return False
+
+                            # The two rooms must not crash if drawn at the same time (size_index > 1.0)
+                            if size_indices[src] + size_indices[dest] >= 1.0:
+                                return False
+
+                            return True
+
+                        # First try each of the unused docks
+                        dest_name = None
+                        dest_dock = None
+                        for (name, dock) in unused_docks:
+                            if are_rooms_compatible(area.name, name):
+                                dest_name = name
+                                dest_dock = dock
+                                break
+
+                        # If that wasn't successful, pick random destinations until it works out
+                        while dest_name is None or dest_dock is None or not are_rooms_compatible(area.name, dest_name):
+                            dest_name = self.rng.choice(world.areas).name
+
+                            if len(area_dock_nums[dest_name]) == 0:
+                                dest_dock = None
+                                continue
+                            
+                            dest_dock = self.rng.choice(area_dock_nums[dest_name])
+
+                        # Don't use this dock as a destination again unless there are no other options
+                        try:
+                            unused_docks.remove((dest_name, dest_dock))
+                        except ValueError:
+                            # print("re-used %s:%d" % (dest_name, dest_dock))
+                            pass
+
+                        world_data[world.name]["rooms"][area.name]["doors"][str(dock_num)] = {
+                            "destination": {
+                                "roomName": dest_name,
+                                "dockNum": dest_dock,
+                            }
+                        }
 
         starting_memo = None
         extra_starting = item_names.additional_starting_items(self.configuration, db.resource_database,
@@ -301,6 +455,9 @@ class PrimePatchDataFactory(BasePatchDataFactory):
             ctwk_config["playerSize"] = 0.3
             ctwk_config["morphBallSize"] = 0.3
             ctwk_config["easyLavaEscape"] = True
+        
+        if self.configuration.large_samus:
+            ctwk_config["playerSize"] = 1.75
 
         if self.cosmetic_patches.use_hud_color:
             ctwk_config["hudColor"] = [
@@ -348,6 +505,29 @@ class PrimePatchDataFactory(BasePatchDataFactory):
 
         seed = self.description.get_seed_for_player(self.players_config.player_index)
 
+        boss_sizes = None
+        if self.configuration.random_boss_sizes:
+            def get_random_size(min, max):
+                if self.rng.choice([True, False]):
+                    return self.rng.uniform(min, 0.8)
+                else:
+                    return self.rng.uniform(1.2, max)
+
+            boss_sizes = {
+                "parasiteQueen": get_random_size(0.1, 3.0),
+                "incineratorDrone": get_random_size(0.2, 3.5),
+                "adultSheegoth": get_random_size(0.2, 1.5),
+                "thardus": get_random_size(0.05, 2.5),
+                "elitePirate1": get_random_size(0.05, 2.3),
+                "elitePirate2": get_random_size(0.05, 1.3),
+                "elitePirate3": get_random_size(0.05, 2.0),
+                "phazonElite": get_random_size(0.05, 2.0),
+                "omegaPirate": get_random_size(0.05, 2.0),
+                "Ridley": get_random_size(0.2, 1.8),
+                "exo": get_random_size(0.05, 2.0),
+                "essence": get_random_size(0.5, 2.25),
+            }
+
         return {
             "seed": seed,
             "preferences": {
@@ -356,7 +536,6 @@ class PrimePatchDataFactory(BasePatchDataFactory):
                 "qolCosmetic": self.cosmetic_patches.qol_cosmetic,
                 "qolPickupScans": self.configuration.qol_pickup_scans,
                 "qolCutscenes": self.configuration.qol_cutscenes.value,
-
                 "mapDefaultState": map_default_state,
                 "artifactHintBehavior": None,
                 "automaticCrashScreen": True,
@@ -366,6 +545,8 @@ class PrimePatchDataFactory(BasePatchDataFactory):
                 "suitColors": suit_colors,
             },
             "gameConfig": {
+                "bossSizes": boss_sizes,
+                "noDoors": self.configuration.no_doors,
                 "shufflePickupPosition": self.configuration.shuffle_item_pos,
                 "shufflePickupPosAllRooms": self.configuration.items_every_room,
                 "startingRoom": starting_room,